+++ sources/components/rb__childfx/rb__childfx.js
@@ -56,14 +56,14 @@
 			 *     }
              * </style>
              *
-             * <div class="rb-main js-rb-life" data-module="childfxExtension">
+             * <div class="rb-main js-rb-live" data-module="childfxExtension">
              *     <img class="logo" src="..." />
              * </div>
              *
              * <script>
              * rb.components._childfx.extend('childfxExtension', {
-			 *      init: function(element){
-			 *          this._super(element);
+			 *      init: function(element, initialDefaults){
+			 *          this._super(element, initialDefaults);
 			 *          this.pos();
 			 *      },
 			 *      pos: function(){