+++ sources/components/rb_itemscroller/rb_itemscroller.js
@@ -35,9 +35,12 @@
              * @property {String}  dragEasing='0.1, 0.25, 0.1, 1.03' Easing value for the slide animation after drag.
              * @property {Number}  duration=600 Average duration for the slide animation.
              * @property {Boolean} mouseDrag=true Whether scroller should be draggable via mouse.
+             * @property {Boolean} wheel=true Whether scroller should be draggable via wheel/trackpad scroll.
              * @property {String|false} dragExclude=false Whether drag/swipe should be excluded on certain element selectors.
              * @property {Boolean} mandatorySnap=false Whether each page generates a mandatory snap point.
              * @property {String}  paginationItemTpl The markup for the pagination buttons.
+             * @property {String|false|undefined}  excludeCell=false Simple selector to exclude cells from scroller cell.
+             * @property {Boolean}  excludeHiddenCells=true Excludes cells that are set to `display: none`.
              * @property {Boolean} switchedOff=false Whether the scroller should be turned off.
              * @property {Boolean} useTransform=true Whether the scroller should use CSS transform3d or left property.
              * @property {Boolean} usePx=false Whether the scroller should use CSS px units instead of % units. Set this to `true` if items are not using % as width unit.
@@ -54,6 +57,7 @@
                 duration: 600,
                 paginationItemTpl: '<span class="{name}{e}pagination{-}btn"></span>',
                 excludeCell: false,
+                excludeHiddenCells: true,
                 useTransform: true,
                 carousel: false,
                 mandatorySnap: false,
@@ -81,7 +85,7 @@
              *
              * @example
              * <!--  markup example -->
-             * <div class="rb-itemscroller js-rb-life" data-module="itemscroller">
+             * <div class="rb-itemscroller js-rb-live" data-module="itemscroller">
              *      {{#if buttons }}
              *          <button type="button" class="itemscroller-btn-prev" tabindex="-1" aria-hidden="true">{{buttons.prev}}</button>
              *          <button type="button" class="itemscroller-btn-next" tabindex="-1" aria-hidden="true">{{buttons.next}}</button>
@@ -178,7 +182,6 @@
                     this.setSwitchedOffClass();
                 }
 
-
             },
             events: {
                 'click:closest(.{name}{e}btn{-}next)': function () {
@@ -246,6 +249,7 @@
                         }
                         break;
                     case 'excludeCell':
+                    case 'excludeHiddenCells':
                         this.updateCells();
                         break;
                 }
@@ -297,20 +301,38 @@
             addWheel: function(){
                 if(!this.options.wheel || !rb.debounce){return;}
 
+                var startValue;
                 var _isWheelStarted = false;
                 var that = this;
                 var options = this.options;
-
+                var block = false;
+                var unblock = function(){
+                    block = false;
+                };
                 var wheelEnd = rb.debounce(function(){
+                    var nearestIndex = that.getNearest();
+                    var diff = that._pos - startValue;
+                    var threshold = Math.min(Math.max(that.viewportWidth / 4, 150), 350);
+
                     _isWheelStarted = false;
-                    that.selectNearest();
-                }, {delay: 40});
+
+                    if(nearestIndex != that._selectedIndex || Math.abs(diff) < threshold){
+                        that.selectedIndex = nearestIndex;
+                    } else if(diff > 0){
+                        that.selectPrev();
+                    } else {
+                        that.selectNext();
+                    }
+                    setTimeout(unblock, 66);
+
+                }, {delay: 44});
 
                 this.viewport.addEventListener('wheel', function(e){
-                    if(!e.deltaMode && !options.switchedOff && options.wheel && Math.abs(e.deltaX) > Math.abs(e.deltaY)){
+                    if(!block && !e.deltaMode && !options.switchedOff && options.wheel && Math.abs(e.deltaX) > Math.abs(e.deltaY)){
 
                         if(!_isWheelStarted){
                             _isWheelStarted = true;
+                            startValue = that._pos;
                             $(that.scroller).stop();
                         }
 
@@ -923,6 +945,13 @@
                 }
 
                 this.$cells = this.$scroller.children(cellSelector);
+
+                if(this.options.excludeHiddenCells){
+                    this.$cells = this.$cells.filter(function(){
+                        return rb.getStyles(this).display != 'none';
+                    });
+                }
+
                 this.calculateLayout();
                 rb.rAFQueue(function () {
                     that.$scroller.prepend(that.helperElem);