+++ sources/js/libs/rb_$.js
@@ -1,7 +1,8 @@
 (function(){
     'use strict';
     if (typeof module == 'object' && module.exports && typeof require != 'undefined') {
-        require('./rb_$/$_slim');
+        var $ = require('./rb_$/$_slim');
         require('./rb_$/$_fx');
+        module.exports = $;
     }
 })();