+++ sources/js/libs/rb_main.js
@@ -192,18 +192,29 @@ if (!window.rb) {
     /* End: rbSlideUp / rbSlideDown */
 
     /* Begin: getScrollingElement */
+
+    //Todo: move into polyfills
+    if(!('scrollingElement' in document)){
+        Object.defineProperty(document, 'scrollingElement', {
+            get: ((document.compatMode == 'BackCompat' || 'WebkitAppearance' in rb.root.style) ?
+                function(){
+                    return  document.body || rb.root;
+                } :
+                function(){
+                    return rb.root;
+                }),
+            enumerable: true,
+            configurable: true,
+        });
+    }
+
     /**
      * @memberof rb
+     * @deprecated use `document.scrollingElement` instead
      * @returns {Element} The DOM element that scrolls the viewport (either html or body)
      */
     rb.getScrollingElement = function () {
-        var scrollingElement = document.scrollingElement;
-
-        if (!scrollingElement && (document.compatMode == 'BackCompat' || 'WebkitAppearance' in rb.root.style)) {
-            scrollingElement = document.body;
-        }
-
-        return scrollingElement || rb.root;
+        return document.scrollingElement;
     };
 
 	/**
@@ -402,29 +413,6 @@ if (!window.rb) {
     };
     /* End: getCSSNumbers */
 
-    /* Begin: camelCase */
-
-    rb.camelCase = (function () {
-        var reg = /-([\da-z])/gi;
-        var camelCase = function (all, found) {
-            return found.toUpperCase();
-        };
-
-        /**
-         * camel cases a String
-         * @alias rb#camelCase
-         * @param str {String} String to camelcase
-         * @returns {String} the camel cased string
-         */
-        var retCamel = function (str) {
-            return str.replace(reg, camelCase);
-        };
-
-        return retCamel;
-    })();
-
-    /* End: camelCase */
-
     /* Begin: memoize */
 
 	/**
@@ -490,9 +478,14 @@ if (!window.rb) {
     /* End: parseValue */
 
     /* Begin: idleCallback */
-    rb.rIC = function(fn){
-        return (window.requestIdleCallback || setTimeout)(fn);
-    };
+    rb.rIC = window.requestIdleCallback ?
+        function(fn){
+            return requestIdleCallback(fn);
+        } :
+        function(fn){
+            return setTimeout(fn);
+        }
+    ;
     /* End: idleCallback */
 
     /* Begin: rAF helpers */
@@ -786,6 +779,7 @@ if (!window.rb) {
                 if(!proxy){
                     proxy = function(e){
                         e.delegatedTarget = e.target.closest(selector);
+                        e.delegateTarget = e.delegatedTarget;
                         if(!e.delegatedTarget){return;}
                         return handler.apply(this, arguments);
                     };
@@ -801,6 +795,7 @@ if (!window.rb) {
                     proxy = function(e){
                         if(e.target.matches(selector)){
                             e.delegatedTarget = e.target;
+                            e.delegateTarget = e.delegatedTarget;
                             return handler.apply(this, arguments);
                         }
                     };
@@ -988,7 +983,7 @@ if (!window.rb) {
                 }
                 parent = parent.parentNode;
             }
-            parent = rb.getScrollingElement();
+            parent = document.scrollingElement;
             scrollableElements.push([parent, parent.scrollTop, parent.scrollLeft]);
         };
 
@@ -1254,7 +1249,7 @@ if (!window.rb) {
 
     var cbs = [];
     var setupClick = function () {
-        var clickClass = ['js', 'click'].join(rb.nameSeparator);
+        var clickClass = ['js', 'rb', 'click'].join(rb.nameSeparator);
         var clickSel = '.' + clickClass;
         var applyBehavior = function (clickElem, e) {
             var i, len, attr, found;
@@ -1302,11 +1297,11 @@ if (!window.rb) {
     };
 
     /**
-     * Allows to add click listeners for fast event delegation. For elements with the class `js-click` and a data-{name} attribute.
+     * Allows to add click listeners for fast event delegation. For elements with the class `js-rb-click` and a data-{name} attribute.
      * @property rb.click.add {Function} add the given name and the function as a delegated click handler.
      * @memberof rb
      * @example
-     * //<a class="js-click" data-lightbox="1"></a>
+     * //<a class="js-rb-click" data-lightbox="1"></a>
      * rb.click.add('lightbox', function(element, event, attrValue){
 	 *
 	 * });
@@ -1468,9 +1463,9 @@ if (!window.rb) {
 (function (window, document) {
     'use strict';
 
-    var elements, useMutationEvents, implicitlyStarted, lifeBatch, initClass, attachedClass, started;
+    var elements, useMutationEvents, implicitlyStarted, liveBatch, initClass, attachedClass, started;
 
-    var life = {};
+    var live = {};
     var removeElements = [];
     var rb = window.rb;
     var $ = rb.$;
@@ -1486,7 +1481,7 @@ if (!window.rb) {
             configurable: true,
             enumerable: true,
             writable: true,
-            value: Object.assign({}, SuperClasss[prop], proto[prop], Class[prop])
+            value: $.extend(true, {}, SuperClasss[prop], proto[prop], Class[prop])
         });
 
         if (proto[prop]) {
@@ -1503,16 +1498,16 @@ if (!window.rb) {
         initClickCreate = $.noop;
         rb.click.add('module', function (elem) {
             if(!rb.getComponent(elem)){
-                rb.log('js-click component not found', elem);
+                rb.log('js-rb-click component not found', elem);
             }
             rb.rAFQueue(function () {
                 elem.classList.remove(rb.click.clickClass);
                 if(!elem.classList.contains(attachedClass)){
                     elem.classList.add(initClass);
-                    life.searchModules();
+                    live.searchModules();
                 }
             }, true);
-            lifeBatch.run();
+            liveBatch.run();
         });
     };
 
@@ -1543,12 +1538,12 @@ if (!window.rb) {
             var main = function () {
                 var len, instance, element;
                 var start = Date.now();
-                for (len = life._attached.length; i < len && Date.now() - start < 6; i++) {
-                    element = life._attached[i];
+                for (len = live._attached.length; i < len && Date.now() - start < 6; i++) {
+                    element = live._attached[i];
 
                     if (element && (instance = element[componentExpando]) && !docElem.contains(element)) {
                         element.classList.add(initClass);
-                        life.destroyComponent(instance, i, element);
+                        live.destroyComponent(instance, i, element);
 
                         i--;
                         len--;
@@ -1581,7 +1576,7 @@ if (!window.rb) {
             for (i = 0; i < len; i++) {
                 mutation = mutations[i];
                 if (mutation.addedNodes.length) {
-                    life.searchModules();
+                    live.searchModules();
                 }
                 if (mutation.removedNodes.length) {
                     removeComponents();
@@ -1594,8 +1589,8 @@ if (!window.rb) {
                 .observe(docElem, {subtree: true, childList: true})
             ;
         } else {
-            docElem.addEventListener('DOMNodeInserted', life.searchModules);
-            document.addEventListener('DOMContentLoaded', life.searchModules);
+            docElem.addEventListener('DOMNodeInserted', live.searchModules);
+            document.addEventListener('DOMContentLoaded', live.searchModules);
             docElem.addEventListener('DOMNodeRemoved', (function () {
                 var mutation = {
                     addedNodes: []
@@ -1651,13 +1646,14 @@ if (!window.rb) {
             });
         }
     };
-
     var mainInit = function(){
+
+        window.removeEventListener('click', mainInit, true);
         mainInit = false;
 
         extendOptions(rb.cssConfig);
 
-        initClass = ['js', 'rb', 'life'].join(rb.nameSeparator);
+        initClass = ['js', 'rb', 'live'].join(rb.nameSeparator);
         attachedClass = ['js', 'rb', 'attached'].join(rb.nameSeparator);
 
         elements = document.getElementsByClassName(initClass);
@@ -1675,25 +1671,25 @@ if (!window.rb) {
 
     rb.ready = rb.deferred();
 
-    rb.life = life;
+    rb.live = live;
 
-    life.autoStart = true;
+    live.autoStart = true;
 
-    life.expando = expando;
-    life.componentExpando = componentExpando;
+    live.expando = expando;
+    live.componentExpando = componentExpando;
 
-    life._failed = {};
+    live._failed = {};
 
     /**
-     * List of all component classes registered by `rb.life.register`.
+     * List of all component classes registered by `rb.live.register`.
      * @memberof rb
      * @type {{}}
      */
     rb.components = {};
-    life._attached = [];
-    life.customElements = false;
+    live._attached = [];
+    live.customElements = false;
 
-    life.init = function (options) {
+    live.init = function (options) {
         if (started) {
             return;
         }
@@ -1706,17 +1702,21 @@ if (!window.rb) {
             extendOptions(options);
         }
 
-        lifeBatch = createBatch();
+        if(mainInit){
+            window.addEventListener('click', mainInit, true);
+        }
 
-        life.searchModules();
+        liveBatch = createBatch();
+
+        live.searchModules();
     };
 
     /**
-     * Registers a component class with a name and manages its lifecycle. An instance of this class will be automatically constructed with the found element as the first argument. If the class has an `attached` or `detached` instance method these methods also will be invoked, if the element is removed or added from/to the DOM. In most cases the given class inherits from [`rb.Component`]{@link rb.Component}. All component classes are added to the `rb.components` namespace.
+     * Registers a component class with a name and manages its livecycle. An instance of this class will be automatically constructed with the found element as the first argument. If the class has an `attached` or `detached` instance method these methods also will be invoked, if the element is removed or added from/to the DOM. In most cases the given class inherits from [`rb.Component`]{@link rb.Component}. All component classes are added to the `rb.components` namespace.
      *
-     * The DOM element/markup of a component class must have a `data-module` attribute with the name as its value. The `data-module` is split by a "/" and only the last part is used as the component name. The part before can be optionally used for [`rb.life.addImportHook`]{@link rb.life.addImportHook}.
+     * The DOM element/markup of a component class must have a `data-module` attribute with the name as its value. The `data-module` is split by a "/" and only the last part is used as the component name. The part before can be optionally used for [`rb.live.addImportHook`]{@link rb.live.addImportHook}.
      *
-     * Usually the element should also have the class `js-rb-life` to make sure it is constructed as soon as it is attached to the document. If the component element has the class `js-click` instead it will be constructed on first click.
+     * Usually the element should also have the class `js-rb-live` to make sure it is constructed as soon as it is attached to the document. If the component element has the class `js-rb-click` instead it will be constructed on first click.
      *
      * @memberof rb
      * @param {String} name The name of your component.
@@ -1729,8 +1729,8 @@ if (!window.rb) {
 	 *
 	 * }
      *
-     * //<button class="js-rb-life" data-module="my-button"></button>
-     * rb.life.register('my-button', MyButton);
+     * //<button class="js-rb-live" data-module="my-button"></button>
+     * rb.live.register('my-button', MyButton);
      *
      * @example
      * class Time {
@@ -1749,11 +1749,11 @@ if (!window.rb) {
 	 *      }
 	 * }
      *
-     * //<span class="js-rb-life" data-module="time"></span>
-     * rb.life.register('time', Time);
+     * //<span class="js-rb-live" data-module="time"></span>
+     * rb.live.register('time', Time);
      *
      */
-    life.register = function (name, Class, _noCheck) {
+    live.register = function (name, Class, _noCheck) {
         var proto = Class.prototype;
         var superProto = Object.getPrototypeOf(proto);
         var superClass = superProto.constructor;
@@ -1781,12 +1781,12 @@ if (!window.rb) {
             if (!started && !implicitlyStarted) {
                 implicitlyStarted = true;
                 setTimeout(function () {
-                    if (!elements && life.autoStart) {
+                    if (!elements && live.autoStart) {
                         $(function () {
                             if (!elements) {
                                 setTimeout(function () {
                                     if (!elements) {
-                                        life.init();
+                                        live.init();
                                     }
                                 });
                             }
@@ -1794,7 +1794,7 @@ if (!window.rb) {
                     }
                 }, 9);
             } else if (elements) {
-                life.searchModules();
+                live.searchModules();
             }
         }
     };
@@ -1811,7 +1811,7 @@ if (!window.rb) {
      *
      * @example
      * //create a short cut.
-     * var addImportHook = rb.life.addImportHook;
+     * var addImportHook = rb.live.addImportHook;
      *
      * addImportHook('slider', function(moduleName, moduleAttributeValue, reject, element){
 	 *      //non-standard futuristic import
@@ -1838,7 +1838,7 @@ if (!window.rb) {
 	 *      require(['./components/' + moduleAttributeValue]);
 	 * });
      */
-    life.addImportHook = function (names, importCallback) {
+    live.addImportHook = function (names, importCallback) {
         var add = function (name) {
             if (unregisteredFoundHook[name]) {
                 rb.log('overrides ' + name + ' import hook', names, importCallback);
@@ -1864,10 +1864,10 @@ if (!window.rb) {
      * @see rb.getComponent
      *
      * @param element
-     * @param LifeClass
+     * @param liveClass
      * @returns {Object}
      */
-    life.create = function (element, LifeClass, initialOpts) {
+    live.create = function (element, liveClass, initialOpts) {
         var instance;
 
         if (mainInit) {
@@ -1875,7 +1875,7 @@ if (!window.rb) {
         }
 
         if (!(instance = element[componentExpando])) {
-            instance = new LifeClass(element, initialOpts);
+            instance = new liveClass(element, initialOpts);
             element[componentExpando] = instance;
         }
 
@@ -1886,17 +1886,17 @@ if (!window.rb) {
         if (!element[expando] && instance && (instance.attached || instance.detached)) {
 
             if ((instance.attached || instance.detached)) {
-                if (life._attached.indexOf(element) == -1) {
-                    life._attached.push(element);
+                if (live._attached.indexOf(element) == -1) {
+                    live._attached.push(element);
                 }
                 if (instance.attached) {
-                    lifeBatch.add(function () {
+                    liveBatch.add(function () {
                         instance.attached();
                     });
                 }
             }
 
-            lifeBatch.timedRun();
+            liveBatch.timedRun();
         }
         element[expando] = true;
         instance._created = true;
@@ -1904,14 +1904,14 @@ if (!window.rb) {
         return instance;
     };
 
-    life.searchModules = (function () {
+    live.searchModules = (function () {
         var removeInitClass = rb.rAF(function () {
             while (removeElements.length) {
                 removeElements.shift().classList.remove(initClass);
             }
         });
         var failed = function (element, id) {
-            life._failed[id] = true;
+            live._failed[id] = true;
             removeElements.push(element);
             rb.log('failed', id, element);
         };
@@ -1943,10 +1943,10 @@ if (!window.rb) {
                 moduleId = moduleId[moduleId.length - 1];
 
                 if (rb.components[moduleId]) {
-                    life.create(element, rb.components[moduleId]);
+                    live.create(element, rb.components[moduleId]);
                     removeElements.push(element);
                 }
-                else if (life._failed[moduleId]) {
+                else if (live._failed[moduleId]) {
                     failed(element, moduleId);
                 }
                 else if ((hook = (unregisteredFoundHook[moduleId] || unregisteredFoundHook['*']))) {
@@ -1966,20 +1966,20 @@ if (!window.rb) {
             }
 
             removeInitClass();
-            lifeBatch.run();
+            liveBatch.run();
         }, {delay: 50});
 
         return findElements;
     })();
 
-    life.destroyComponent = function (instance, index, element) {
+    live.destroyComponent = function (instance, index, element) {
 
         if (!element) {
             element = instance.element;
         }
 
         if (index == null) {
-            index = life._attached.indexOf(element);
+            index = live._attached.indexOf(element);
         }
         element.classList.remove(attachedClass);
 
@@ -1990,10 +1990,10 @@ if (!window.rb) {
             instance.detached(element, instance);
         }
 
-        life._attached.splice(index, 1);
+        live._attached.splice(index, 1);
     };
 
-    return life;
+    return live;
 })(window, document);
 
 /* Simple JavaScript Inheritance
@@ -2098,12 +2098,13 @@ if (!window.rb) {
     rb.Class = ES5Class;
 })();
 
-(function (window, document, life, undefined) {
+(function (window, document, live, undefined) {
     'use strict';
 
+    var focusClass, focusSel;
     var rb = window.rb;
     var $ = rb.$;
-    var componentExpando = life.componentExpando;
+    var componentExpando = live.componentExpando;
     var regData = /^data-/;
     var regWhite = /\s+(?=[^\)]*(?:\(|$))/g;
     var regComma = /\s*,\s*(?=[^\)]*(?:\(|$))/g;
@@ -2157,6 +2158,12 @@ if (!window.rb) {
             return str.replace(regHTMLSel, replacer);
         };
     })(), true);
+    var generateFocusClasses = function(){
+        focusClass = ['js', 'rb', 'autofocus'].join(rb.nameSeparator);
+        focusSel = '.' + focusClass;
+    };
+
+    rb.ready.then(generateFocusClasses);
 
     rb.parseEventString = rb.memoize(function(evtStr){
 
@@ -2201,7 +2208,7 @@ if (!window.rb) {
     rb.getComponent = function (element, componentName, initialOpts) {
         var component = element && element[componentExpando];
 
-        if (!component) {
+        if (!component && element) {
 
             if (!rb.components[componentName]) {
                 componentName = (element.getAttribute('data-module') || '').split('/');
@@ -2210,13 +2217,13 @@ if (!window.rb) {
             }
 
             if (rb.components[componentName]) {
-                component = life.create(element, rb.components[componentName], initialOpts);
+                component = live.create(element, rb.components[componentName], initialOpts);
             }
         }
-        return component;
+        return component || null;
     };
 
-    life.getComponent = rb.getComponent;
+    live.getComponent = rb.getComponent;
 
     rb.Component = rb.Class.extend(
         /** @lends rb.Component.prototype */
@@ -2225,13 +2232,13 @@ if (!window.rb) {
              * constructs the class
              * @classdesc Component Base Class - all UI components should extend this class. This Class adds some neat stuff to parse/change options (is automatically done in constructor), listen and trigger events, react to responsive state changes and establish BEM style classes as also a11y focus management.
              *
-             * For the life cycle features see [rb.life.register]{@link rb.life.register}.
+             * For the live cycle features see [rb.live.register]{@link rb.live.register}.
              * @param element
              * @param [initialDefaults] {Object}
              * @constructs
              *
              * @example
-             * //<div class="js-rb-life" data-module="my-module"></div>
+             * //<div class="js-rb-live" data-module="my-module"></div>
              * rb.Component.extend('my-component', {
 			 *      defaults: {
 			 *          className: 'toggle-class',
@@ -2248,11 +2255,11 @@ if (!window.rb) {
 			 *          this.$element.toggleClass(this.options.className);
 			 *      }
 			 * });
-             * //If ES5 static extend method is used, rb.life.register is called implicitly.
+             * //If ES5 static extend method is used, rb.live.register is called implicitly.
              *
              * @example
-             * //<div class="js-rb-life" data-module="my-module"></div>
-             * rb.life.register('my-component', class MyComponent extends rb.Component {
+             * //<div class="js-rb-live" data-module="my-module"></div>
+             * rb.live.register('my-component', class MyComponent extends rb.Component {
 			 *      static get defaults(){
 			 *          return {
 			 *              className: 'toggle-class',
@@ -2334,7 +2341,7 @@ if (!window.rb) {
              * @example
              *
              * //creating a new component with different defaults:
-             * rb.life.register('special-accordion', class SpecialAccordion extends rb.components.accordion {
+             * rb.live.register('special-accordion', class SpecialAccordion extends rb.components.accordion {
 			 *      static get defaults(){
 			 *          return {
 			 *              multiple: true,
@@ -2385,8 +2392,8 @@ if (!window.rb) {
              * @example
              *
              * class MyComponent extends rb.Component {
-			 *      constructor(element){
-			 *          super(element);
+			 *      constructor(element, initialDefaults){
+			 *          super(element, initialDefaults);
 			 *          this.child = this.query('.{name}__close-button');
 			 *
 			 *          rb.rAFs(this, 'setLayout');
@@ -2409,7 +2416,12 @@ if (!window.rb) {
              * Shortcut to [`rb.getComponent`]{@link rb.getComponent}
              * @function
              */
-            component: rb.getComponent,
+            component: function(element, name, initialOpts){
+                if(typeof element == 'string'){
+                    element = this.query(element);
+                }
+                return rb.getComponent(element, name, initialOpts);
+            },
 
             /**
              * returns the id of an element, if no id exist generates one for the element
@@ -2493,17 +2505,17 @@ if (!window.rb) {
                         focusElement = rb.elementFromStr(element, this.element)[0];
                     }
                 } else {
-                    focusElement = this.element.querySelector('.js-autofocus');
+                    focusElement = this.query(focusSel);
                 }
 
-                if (!focusElement && (element === true || this.element.classList.contains('js-autofocus'))) {
+                if (!focusElement && (element === true || this.element.classList.contains(focusClass))) {
                     focusElement = this.element;
                 }
                 return focusElement;
             },
 
             /**
-             * Sets the focus and remembers the activeElement before. If setComponentFocus is invoked with no argument. The element with the class `js-autofocus` inside of the component element is focused.
+             * Sets the focus and remembers the activeElement before. If setComponentFocus is invoked with no argument. The element with the class `js-rb-autofocus` inside of the component element is focused.
              * @param [element] {Element|Boolean|String} The element that should be focused. In case a string is passed the string is converted to an element using `rb.elementFromStr`
              * @param [delay] {Number} The delay that should be used to focus an element.
              */
@@ -2633,7 +2645,7 @@ if (!window.rb) {
              * @param initialOpts
              */
             parseOptions: function (opts) {
-                var options = Object.assign(opts || {}, this.constructor.defaults, this._initialDefaults, this.parseCSSOptions(), this.parseHTMLOptions());
+                var options = $.extend(true, opts || {}, this.constructor.defaults, this._initialDefaults, this.parseCSSOptions(), this.parseHTMLOptions());
                 this.setOptions(options);
             },
 
@@ -2642,9 +2654,17 @@ if (!window.rb) {
              * @param opts
              */
             setOptions: function (opts) {
-                for (var prop in opts) {
-                    if (opts[prop] !== this.options[prop]) {
-                        this.setOption(prop, opts[prop]);
+                var oldValue, newValue;
+
+                if(opts !== this.options){
+                    for (var prop in opts) {
+                        newValue = opts[prop];
+                        oldValue = this.options[prop];
+                        if (newValue !== oldValue &&
+                            (!oldValue || typeof newValue != 'object' || typeof oldValue != 'object' ||
+                            JSON.stringify(newValue) != JSON.stringify(oldValue))) {
+                            this.setOption(prop, newValue);
+                        }
                     }
                 }
             },
@@ -2708,7 +2728,7 @@ if (!window.rb) {
              *
              * //require('sources/js/_templates/my-component.js');
              *
-             * rb.life.register('my-component', class MyComponent extends rb.Component {
+             * rb.live.register('my-component', class MyComponent extends rb.Component {
 			 *      renderMain(){
 			 *          this.element.innerHTML = this.render('main', {title: 'foo'});
 			 *      }
@@ -2750,7 +2770,7 @@ if (!window.rb) {
                 for (i = 0; i < len; i++) {
                     name = attributes[i].nodeName;
                     if (name != optionsAttr && name.startsWith('data-') && !name.endsWith('-options')) {
-                        options[rb.camelCase(name.replace(regData, ''))] = rb.parseValue(attributes[i].nodeValue);
+                        options[$.camelCase(name.replace(regData, ''))] = rb.parseValue(attributes[i].nodeValue);
                     }
                 }
 
@@ -2767,7 +2787,7 @@ if (!window.rb) {
             },
 
             destroy: function () {
-                life.destroyComponent(this);
+                live.destroyComponent(this);
             },
 
             _super: function () {
@@ -2793,7 +2813,7 @@ if (!window.rb) {
             Class.prototype.statics = null;
         }
 
-        life.register(name, Class, noCheck);
+        live.register(name, Class, noCheck);
         return Class;
     };
 
@@ -2813,7 +2833,9 @@ if (!window.rb) {
         return Class;
     };
 
-})(window, document, rb.life);
+    generateFocusClasses();
+
+})(window, document, rb.live);
 
 (function (window, document, undefined) {
     'use strict';
@@ -2846,7 +2868,7 @@ if (!window.rb) {
              * ```html
              * <button type="button"
              *  data-module="button"
-             *  class="js-click"
+             *  class="js-rb-click"
              *  aria-controls="panel-1"
              *  data-type="open">
              *      click me