+++ CHANGELOG.md
@@ -2,5 +2,12 @@
 
 This changelog only includes breaking change information.
 
-#upcomming 0.3.0
+#0.4.2
+    * rename js-autofocus to js-rb-autofocus
+
+#0.4.0
+    * renamed nearly all uses of `life` to `live`
+    * renamed `js-click` to `js-rb-click`
+
+#0.3.0
     * renamed `rbfocusenter`/`rbfocusleave` to `rb_focusenter`/`rb_focusleave`